fix(pci-instance): distant backup - react - extract component/hook from BackupAction page
ref: #TAPC-5162
